Split model page specs contracts
